pyproject.toml
[project]
name = "blog-post"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "generics>=7.0.0",
    "pillow>=11.3.0",
]
